change the declaration of "lineno" at line 533
svn path=/trunk/KDE/kdelibs/; revision=1043933

+++ khtml/xml/xml_tokenizer.cpp
@@ -533,7 +533,7 @@ void XMLTokenizer::finish()
         if ( m_handler.errorLine != -1 ) {
             QString xmlCode = m_source.data();
             QTextStream stream(&xmlCode, QIODevice::ReadOnly);
-            for (unsigned long lineno = 0; lineno < m_handler.errorLine-1; lineno++)
+            for (int lineno = 0; lineno < m_handler.errorLine-1; lineno++)
                 stream.readLine();
             line = stream.readLine();
